Add support for special theme for unfocused outputs
I use multiple outputs, and as such often need to remember which
output is currently focused.

Fortunately, river tells us which output is currently in focus.
Therefore, we can track this in river's info provider, and then use
this value to decide which theme to use when drawing the bar.

For users that do not need or want this extra differentiation, setting
the `[theme.palette]` values, will result in the previous behaviour as
the `[theme.unfocused_palette]` will inherit them if not set.

By default, that is with `[theme.palette]` unset, we will use a darker
white for the `unfocused_palette`'s foreground, to signify that this
output is unfocused.

I am not familiar enough with niri or hyprland to know if they can
support something like this too. As such, both of their information
providers do not override the default `is_output_focused` method.

In the future moving even more values to the themes might be a viable
approach (i.e., the font), but with the current values most workflows
should be possible (e.g., hiding inactive tags on the unfocused output,
whilst showing them on the focused one or simply setting darker colours).