Show the user the concrete configuration error
The `Display` impl on anyhows's errors will just show the topmost error.
But the user is probably interested in what exactly was wrong with their
config, not only that it was wrong. So we now simply print the full
error to stderr, which will include a nice description of what serde
expected.